+++ Main.rb
@@ -1,9 +1,9 @@
 # Change filename to Main.rb
-require './CardValidator'
+require './card_validator'
 puts "Enter card numbers: "
 while true
-    cardNumber = gets.chomp.delete(" ")
-    card = Card.new(cardNumber)
-    card.validate
-    puts "#{card.type}: #{cardNumber} #{card.valid}" 
+  cardNumber = gets.chomp.delete(" ")
+  card = CardValidator.new(cardNumber)
+  card.validate
+  puts "#{card.type}: #{cardNumber} #{card.valid}" 
 end
